requirements/dev.txt
-r base.txt
ipython==8.29.0
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
ruff==0.8.0
